add testing for the types and fix time millis to use int. Also protect row writer

+++ hudi-common/src/main/java/org/apache/hudi/internal/schema/Type.java
@@ -62,7 +62,7 @@ public interface Type extends Serializable {
     DATE(Integer.class),
     BOOLEAN(Boolean.class),
     TIME(Long.class),
-    TIME_MILLIS(Long.class),
+    TIME_MILLIS(Integer.class),
     TIMESTAMP(Long.class),
     TIMESTAMP_MILLIS(Long.class),
     DECIMAL(BigDecimal.class),
